extract SBOM output structs; share across tests
